Create TM Trainer classes
