+++ tests/Neat_Trainer_save_inputs.py
@@ -1,25 +1,80 @@
 """
-Utils for TrackManAI
+Classes to get the NEAT model running.
+There are classes to :
+    - Asynchronously capture screens of a window and access the captured screen.
+    - Perform a generation step
 """
 
 # Imports
-import numpy as np
-import win32gui
-import win32con
-import win32ui
+from encodings import normalize_encoding
+from xmlrpc.client import Boolean
 from tminterface.client import Client
 from tminterface.interface import TMInterface
 from tminterface.constants import DEFAULT_SERVER_SIZE
 import signal
+from PIL import ImageGrab, ImageEnhance, Image
+import keyboard
+import sys
+import numpy as np
+import neat
 import time
+import os
+import cv2
+import pickle as pickle
 
+#os.environ["PATH"] = r"d:D:\ProgramData\Anaconda3\envs\trackmanAIenv\Lib\site-packagespywin32_system32;" + os.environ["PATH"]
+
+import win32gui
+import win32con
+import win32ui
+from threading import Thread, Lock
+
+# In order to visualize the training net, you need to copy visualize.py file into the NEAT directory (you can find it in the NEAT repo)
+# Because of the licence, I am not going to copy it to my github repository
+# You can still train your network without it
+try:
+    import neat.visualize as visualize
+except ModuleNotFoundError:
+    print('Missing visualize.py file.')
+
+#os.chdir('./NEAT')
+#print(os.getcwd())
+
+if len(sys.argv) < 0:
+    print('Not enough arguments.')
+    exit()
+
+
+# Parameters : these should be put in a config file
+# image dimensions
+w=1280
+h=960
+
+n_lines=20
+# hyperparams
+#threshold = 0.5
+no_generations = 1000 #20 for straight with 8sec of max
+#max_fitness = 100.0
+gamespeed=1
+skip_frames=4
+kill_time= 3
+kill_speed = 10
+max_time=12
+no_lines = 20 #need to investigate to upscale that
+filename_prefix = "models/NEAT/Checkpoints/checkpoint-"
+checkpoint = filename_prefix+"231" #None # filename_prefix + "neat-checkpoint-0"
+gen=231 #current gen
+server_name=f'TMInterface{sys.argv[1]}' if len(sys.argv)>1 else 'TMInterface0'
+
+
+
+#from https://nicholastsmith.wordpress.com/2017/08/10/poe-ai-part-4-real-time-screen-capture-and-plumbing/?fbclid=IwAR3ZHfVY2oPr1kqhq_o4EthijXh1GLDoK2FYw3bWReRWMUEBWTB8_jhwd1Q
 
 # Classes
 class ScreenViewer:
     """ Asynchronously captures screens of a window. Provides functions for accessing the captured screen.
-    (from https://nicholastsmith.wordpress.com/2017/08/10/poe-ai-part-4-real-time-screen-capture-and-plumbing/?fbclid=IwAR3ZHfVY2oPr1kqhq_o4EthijXh1GLDoK2FYw3bWReRWMUEBWTB8_jhwd1Q)
     """
-    def __init__(self, n_lines, w, h):
+    def __init__(self):
         #self.mut = Lock()
         self.hwnd = None
         self.its = None         #Time stamp of last image 
@@ -31,19 +86,16 @@ class ScreenViewer:
         #Border on left and top to remove
         #self.bl, self.bt, self.br, self.bb = 12, 31, 12, 20
         self.bl, self.bt, self.br, self.bb = 0, 0, 0, 0
-        self.L_pix_lines=get_pix_lines(n_lines)
-        self.w = w
-        self.h = h
+        self.L_pix_lines=Get_pix_lines(n_lines)
     
 
-    def getHWND(self, wname):
+    def GetHWND(self, wname):
         """ Gets handle of window to view.
         Also updates class parameters for the screen window on success.
     
         Parameters
         ----------
         wname: str (Title of window to find)
-
         Output
         ----------
         out : bool (True on success; False on failure)
@@ -56,13 +108,12 @@ class ScreenViewer:
         return True
          
 
-    def getScreenImg(self):
+    def GetScreenImg(self):
         """ Gets the screen of the window referenced by self.hwnd
     
         Parameters
         ----------
         None
-
         Output
         ----------
         im : np.ndarray (screen image)
@@ -80,12 +131,12 @@ class ScreenViewer:
         dcObj = win32ui.CreateDCFromHandle(wDC)
         cDC = dcObj.CreateCompatibleDC()
         dataBitMap = win32ui.CreateBitmap()
-        dataBitMap.CreateCompatibleBitmap(dcObj, self.w, self.h)
+        dataBitMap.CreateCompatibleBitmap(dcObj, w, h)
         cDC.SelectObject(dataBitMap)
         #First 2 tuples are top-left and bottom-right of destination
         #Third tuple is the start position in source
         #cDC.BitBlt((0,0), (w, h), dcObj, (self.bl, self.bt), win32con.SRCCOPY)
-        cDC.BitBlt((0,0), (self.w, self.h), dcObj, (0, 38), win32con.SRCCOPY)
+        cDC.BitBlt((0,0), (w, h), dcObj, (0, 38), win32con.SRCCOPY)
         #dataBitMap.SaveBitmapFile(cDC, 'debug.bmp')
         bmInfo = dataBitMap.GetInfo()
         im = np.frombuffer(dataBitMap.GetBitmapBits(True), dtype = np.uint8)
@@ -102,10 +153,139 @@ class ScreenViewer:
         return im
 
 
+# Utils functions (should be put in an utils file)
+def get_end_points(n_rays):
+    """ Gets the end points for a given number of rays
+    Parameters
+    ----------
+    n_rays: int (number of rays to consider)
+    Output
+    ----------
+    start_point: (int, int) (start point for rays propagation)
+    L_points : Array([int, int]) (coords of every end point)
+    """
+    shape=(1279,959)
+    start_point=(int(shape[0]/2),shape[1]-60) #OK
+    L_teta=[ np.pi*i/(n_rays-1) for i in range(n_rays)] #OK
+    ll=((start_point[0]-0)**2 + (start_point[1]-0)**2)**0.5
+    lr=((start_point[0]-1279)**2 + (start_point[1]-0)**2)**0.5
+    tetal=np.arccos(-(start_point[0]-0)/ll)#ok
+    tetar=np.arccos(-(start_point[0]-1279)/lr)#ok
+    L_end_points=[]
+    for teta in L_teta:
+        if teta<=tetar:
+            x=1279
+            if x-start_point[0]!=0:
+                l=abs((x-start_point[0])/np.cos(teta))
+                y=-l*np.sin(teta)+start_point[1]
+            else:
+                y=start_point[1]
+        elif teta>=tetal:
+            x=0
+            if x-start_point[0]!=0:
+                l=abs((x-start_point[0])/np.cos(teta))
+                y=-l*np.sin(teta)+start_point[1]
+            else:
+                y=start_point[1]
+        else:
+            y=0
+            if abs(y-start_point[1])>1:
+                l=abs((y-start_point[1])/np.sin(teta))
+                x=l*np.cos(teta)+start_point[0]
+            else:
+                x=start_point[0]
+        L_end_points.append([x,y])
+    return start_point,L_end_points
+
+
+#based from https://stackoverflow.com/questions/25837544/get-all-points-of-a-straight-line-in-python?fbclid=IwAR2y-tW6Qmk_1I28KQRF2WslyfmXAFhlQ3_2l0tKL8RQ7qAIj-f6QgBE-NM
+def getLine(x1,y1,x2,y2): #seems good
+    """ Compute a straight line from (x1, y1) to (x2, y2)
+    Parameters
+    ----------
+    x1: int
+    y1: int
+    x2: int
+    y2: int
+    Output
+    ----------
+    line: Array([int, int]) (each point from (x1, y1) to (x2, y2))
+    """
+    if x1==x2: ## Perfectly horizontal line, can be solved easily
+        return [[int(x1),int(i)] for i in range(y1,y2,int(abs(y2-y1)/(y2-y1)))]
+    else: ## More of a problem, ratios can be used instead
+        x_inv=False
+        if x1>x2: ## If the line goes "backwards", flip the positions, to go "forwards" down it.
+            x=x1
+            x1=x2
+            x2=x
+            y=y1
+            y1=y2
+            y2=y
+            x_inv=True
+        slope=(y2-y1)/(x2-x1) ## Calculate the slope of the line
+        line=[]
+        i=0
+        while x1+i < x2: ## Keep iterating until the end of the line is reached
+            i+=1
+            x_end=x1+i
+            y_end=y1+slope*i
+            line.append([int(x_end),int(y_end)]) ## Add the next point on the line
+        if x_inv:
+            line.reverse()
+        return line ## Finally, return the line!
+
+
+def intersect(im,line):
+    """ Gets the intersection index between the line and a different color object (i.e. a wall)
+    Parameters
+    ----------
+    im: np.ndarray (screen image)
+    line: Array([int, int])
+    Output
+    ----------
+    i: int (index on the line that intersects a different object)
+    """
+    for i in range(len(line)):
+        pix=line[i]
+        shape=im.shape
+        #print(shape)
+        if pix[0]<shape[1] and pix[1]<shape[0]:
+            color=sum(im[pix[1]][pix[0]])/len(im[pix[1]][pix[0]])
+            #print(im[pix[0]][pix[1]])
+            #print(color)
+            if color<30:
+                return i
+        else:
+            print(shape)
+            print(pix)
+    return len(line)
+
+
+def Get_Raycast(im,L_pix_lines):
+    """ Gets every (corrected) raycasts on the image
+    """
+    L_intersect_normed=[]
+    for i in range(len(L_pix_lines)):
+        inter=intersect(im,L_pix_lines[i])
+        L_intersect_normed.append(2*inter/len(L_pix_lines[i])-1)
+    return L_intersect_normed
+
+
+def Get_pix_lines(n_lines):
+    """ Gets every raycasts on the image
+    """
+    c,L_end_points=get_end_points(n_lines)
+    L_pix_lines=[]
+    for i in range(len(L_end_points)):
+        L_pix_lines.append(getLine(c[0],c[1],L_end_points[i][0],L_end_points[i][1]))
+    return L_pix_lines
+
+
 class GenClient(Client):
     """ Class to perform a generation step using tminterface.
     """
-    def __init__(self, L_net, max_time, kill_time, sv, skip_frames, wn, gs, ks):
+    def __init__(self,L_net,max_time,kill_time):
         super(GenClient,self).__init__()
         self.init_step=True
         self.fitness=0
@@ -126,10 +306,8 @@ class GenClient(Client):
         self.ready_max_steps=60
         self.ready_current_steps=0
         self.finished=False
-        self.sv=sv
-        self.sv.getHWND(wn)
-        self.gamespeed = gs
-        self.kill_speed = ks
+        self.sv=ScreenViewer()
+        self.sv.GetHWND('TrackMania United Forever (TMInterface 1.2.0)')
 
 
     def on_registered(self, iface: TMInterface):
@@ -139,8 +317,8 @@ class GenClient(Client):
         #iface.log("Ready. Genome id: " + str(self.genome_id))
         #set gamespeed
         iface.set_timeout(5000)
-        if self.gamespeed!=1.0:
-            iface.set_speed(self.gamespeed)
+        if gamespeed!=1.0:
+            iface.set_speed(gamespeed)
         iface.give_up()
     
 
@@ -162,12 +340,10 @@ class GenClient(Client):
     def on_run_step(self, iface, _time:int):
         """ Called on each “run” step (physics tick). 
         This method will be called only in normal races and not when validating a replay.
-
         Parameters
         ----------
         iface: TMInterface (the TMInterface object)
         _time: int (the physics tick)
-
         Output
         ----------
         None
@@ -198,14 +374,14 @@ class GenClient(Client):
                 # Initialize step
                 self.init_step=False
                 self.current_step+=1
-                self.im=self.sv.getScreenImg() #try
+                self.im=self.sv.GetScreenImg() #try
                 #self.img=Image.fromarray(self.img,"RGB")
                 #self.img.save(f"{self.time}".zfill(10)+".png")
                 #self.img = ImageGrab.grab()
                 #Image.new(self.img).save(f"{self.time}".zfill(10)+".png")
                 self.L_pix_lines=self.sv.L_pix_lines
                 #self.img = ImageGrab.grab()
-                self.L_raycast = get_raycast(self.im,self.L_pix_lines)
+                self.L_raycast = Get_Raycast(self.im,self.L_pix_lines)
                 self.inputs=self.L_raycast
                 self.inputs.append(speed)
                 self.inputs.append(self.yaw)
@@ -223,10 +399,10 @@ class GenClient(Client):
                 # Update step
                 #one screenshot every skip_frame frames
                 if self.current_step%self.skip_frames==0:
-                    self.im=self.sv.getScreenImg()
+                    self.im=self.sv.GetScreenImg()
                     #self.img.save("test.png") #to work on the frame processing
                     self.L_pix_lines=self.sv.L_pix_lines
-                    self.L_raycast = get_raycast(self.im,self.L_pix_lines)
+                    self.L_raycast = Get_Raycast(self.im,self.L_pix_lines)
                     self.inputs=self.L_raycast
                     self.inputs.append(speed)
                     self.inputs.append(self.yaw)
@@ -265,7 +441,7 @@ class GenClient(Client):
                         iface.close()
                         self.finished=True
                 else:
-                    if self.time>self.kill_time and self.speed<self.kill_speed:
+                    if self.time>self.kill_time and self.speed<kill_speed:
                         #iface.log(str(self.speed))
                         #iface.log(str(self.fitness))
                         
@@ -306,13 +482,11 @@ class GenClient(Client):
     def on_checkpoint_count_changed(self, iface, current:int, target:int):
         """ Called when the current checkpoint count changed 
         (a new checkpoint has been passed by the vehicle).
-
         Parameters
         ----------
         iface: TMInterface (the TMInterface object)
         current: int (the current amount of checkpoints passed)
         target: int (the total amount of checkpoints on the map (including finish))
-
         Output
         ----------
         None
@@ -355,7 +529,6 @@ class GenClient(Client):
     def on_custom_command(self, iface, time_from: int, time_to: int, command: str, args: list):
         """
         Called when a custom command has been executed by the user.
-
         Parameters
         ----------
         iface: TMInterface (the TMInterface object)
@@ -363,7 +536,6 @@ class GenClient(Client):
         time_to: int (if provided by the user, the ending time of the command, otherwise -1)
         command: str (the command name being executed)
         args: list (the argument list provided by the user)
-
         Output
         ----------
         None
@@ -375,184 +547,16 @@ class GenClient(Client):
         """
         Called when a client exception is thrown. This can happen if opening the shared file fails, or reading from
         it fails.
-
         Parameters
         ----------
         iface: TMInterface (the TMInterface object)
         exception: Exception (the exception being thrown)
-
         Output
         ----------
         None
         """
         print(f'[Client] Exception reported: {exception}')
         #iface.register(self)#try
-
-
-class TMTrainer:
-    """ TrackManAI Trainer.
-    """
-    def __init__(self, model_dir, model_config, w, h, window_name, 
-    n_lines, server_name, gamespeed, skip_frames, 
-    kill_time, kill_speed, max_time, no_lines,
-    screen_viewer, checkpoint=None):
-        self.model_dir = model_dir
-        self.model_config = model_config
-        self.w = w
-        self.h = h
-        self.window_name = window_name
-        self.n_lines = n_lines
-        self.server_name = server_name
-        self.gamespeed = gamespeed
-        self.skip_frames = skip_frames
-        self.kill_time = kill_time
-        self.kill_speed = kill_speed
-        self.max_time = max_time
-        self.no_lines = no_lines
-        self.screen_viewer = screen_viewer
-        self.checkpoint = checkpoint
-        if checkpoint == None:
-            self.filename_prefix = model_dir
-            self.gen = 0
-        else:
-            checkpoint_infos = checkpoint.split('/')[-1].split('-')
-            self.filename_prefix = checkpoint_infos[0]
-            self.gen = int(checkpoint_infos[-1])
-
-
-# Functions
-def get_end_points(n_rays):
-    """ Gets the end points for a given number of rays
-
-    Parameters
-    ----------
-    n_rays: int (number of rays to consider)
-
-    Output
-    ----------
-    start_point: (int, int) (start point for rays propagation)
-    L_points : Array([int, int]) (coords of every end point)
-    """
-    shape=(1279,959)
-    start_point=(int(shape[0]/2),shape[1]-60) #OK
-    L_teta=[ np.pi*i/(n_rays-1) for i in range(n_rays)] #OK
-    ll=((start_point[0]-0)**2 + (start_point[1]-0)**2)**0.5
-    lr=((start_point[0]-1279)**2 + (start_point[1]-0)**2)**0.5
-    tetal=np.arccos(-(start_point[0]-0)/ll)#ok
-    tetar=np.arccos(-(start_point[0]-1279)/lr)#ok
-    L_end_points=[]
-    for teta in L_teta:
-        if teta<=tetar:
-            x=1279
-            if x-start_point[0]!=0:
-                l=abs((x-start_point[0])/np.cos(teta))
-                y=-l*np.sin(teta)+start_point[1]
-            else:
-                y=start_point[1]
-        elif teta>=tetal:
-            x=0
-            if x-start_point[0]!=0:
-                l=abs((x-start_point[0])/np.cos(teta))
-                y=-l*np.sin(teta)+start_point[1]
-            else:
-                y=start_point[1]
-        else:
-            y=0
-            if abs(y-start_point[1])>1:
-                l=abs((y-start_point[1])/np.sin(teta))
-                x=l*np.cos(teta)+start_point[0]
-            else:
-                x=start_point[0]
-        L_end_points.append([x,y])
-    return start_point,L_end_points
-
-
-def getLine(x1,y1,x2,y2):
-    """ Compute a straight line from (x1, y1) to (x2, y2)
-    (based on https://stackoverflow.com/questions/25837544/get-all-points-of-a-straight-line-in-python?fbclid=IwAR2y-tW6Qmk_1I28KQRF2WslyfmXAFhlQ3_2l0tKL8RQ7qAIj-f6QgBE-NM)
-
-    Parameters
-    ----------
-    x1: int
-    y1: int
-    x2: int
-    y2: int
-
-    Output
-    ----------
-    line: Array([int, int]) (each point from (x1, y1) to (x2, y2))
-    """
-    if x1==x2: ## Perfectly horizontal line, can be solved easily
-        return [[int(x1),int(i)] for i in range(y1,y2,int(abs(y2-y1)/(y2-y1)))]
-    else: ## More of a problem, ratios can be used instead
-        x_inv=False
-        if x1>x2: ## If the line goes "backwards", flip the positions, to go "forwards" down it.
-            x=x1
-            x1=x2
-            x2=x
-            y=y1
-            y1=y2
-            y2=y
-            x_inv=True
-        slope=(y2-y1)/(x2-x1) ## Calculate the slope of the line
-        line=[]
-        i=0
-        while x1+i < x2: ## Keep iterating until the end of the line is reached
-            i+=1
-            x_end=x1+i
-            y_end=y1+slope*i
-            line.append([int(x_end),int(y_end)]) ## Add the next point on the line
-        if x_inv:
-            line.reverse()
-        return line ## Finally, return the line!
-
-
-def intersect(im,line):
-    """ Gets the intersection index between the line and a different color object (i.e. a wall)
-
-    Parameters
-    ----------
-    im: np.ndarray (screen image)
-    line: Array([int, int])
-
-    Output
-    ----------
-    i: int (index on the line that intersects a different object)
-    """
-    for i in range(len(line)):
-        pix=line[i]
-        shape=im.shape
-        #print(shape)
-        if pix[0]<shape[1] and pix[1]<shape[0]:
-            color=sum(im[pix[1]][pix[0]])/len(im[pix[1]][pix[0]])
-            #print(im[pix[0]][pix[1]])
-            #print(color)
-            if color<30:
-                return i
-        else:
-            print(shape)
-            print(pix)
-    return len(line)
-
-
-def get_raycast(im,L_pix_lines):
-    """ Gets every (corrected) raycasts on the image
-    """
-    L_intersect_normed=[]
-    for i in range(len(L_pix_lines)):
-        inter=intersect(im,L_pix_lines[i])
-        L_intersect_normed.append(2*inter/len(L_pix_lines[i])-1)
-    return L_intersect_normed
-
-
-def get_pix_lines(n_lines):
-    """ Gets every raycasts on the image
-    """
-    c,L_end_points=get_end_points(n_lines)
-    L_pix_lines=[]
-    for i in range(len(L_end_points)):
-        L_pix_lines.append(getLine(c[0],c[1],L_end_points[i][0],L_end_points[i][1]))
-    return L_pix_lines
 
 
 def run_client_gen(client: Client, server_name: str = 'TMInterface0', buffer_size=DEFAULT_SERVER_SIZE):
@@ -562,13 +566,11 @@ def run_client_gen(client: Client, server_name: str = 'TMInterface0', buffer_siz
     until the client is deregistered in any way. You can set the buffer size yourself to use for
     the connection, by specifying the buffer_size parameter. Using a custom size requires
     launching TMInterface with the /serversize command line parameter: TMInterface.exe /serversize=size.
-
     Parameters
     ----------
     client: Client (the client instance to register)
     server_name: str (the server name to connect to, TMInterface0 by default)
     buffer_size: int (the buffer size to use, the default size is defined by tminterface.constants.DEFAULT_SERVER_SIZE)
-
     Output
     ----------
     L_fit: Array(int) (Fitness for each NN that ran a simulation)
@@ -593,3 +595,128 @@ def run_client_gen(client: Client, server_name: str = 'TMInterface0', buffer_siz
     iface.close()
 
     return client.L_fit,client.L_coords,client.L_speeds,client.L_inputs
+
+
+def eval_genomes(genomes, config):
+    """
+    Evaluates every genome (NN) of the generation
+    Parameters
+    ----------
+    genomes: Array(genome) (every NN of the simulation)
+    config: neat.Config (config for the neat algorithm)
+    Output
+    ----------
+    None
+    """
+    # Initialize client
+    global gen,kill_time,max_time
+    gen+=1
+    L_net=[]
+    for genome_id, genome in genomes:
+        net = neat.nn.FeedForwardNetwork.create(genome, config)
+        L_net.append(net)
+    max_time2=min(max_time,1+0.5*gen)
+
+    # Run gen
+    L_fit,L_coords,L_speeds,L_inputs=run_client_gen(GenClient(L_net,max_time2,kill_time),server_name) #1/6 de sec en plus par génération
+
+    # Update fitness
+    for i in range(len(L_fit)):
+        genomes[i][1].fitness=L_fit[i]
+
+    # Write generation recap
+    filename = 'models/NEAT/Coords/'+str(gen).zfill(5)+'.pickle'
+    outfile = open(filename,'wb')
+    pickle.dump(L_coords,outfile)
+    outfile.close()
+    filename = 'models/NEAT/Speeds/'+str(gen).zfill(5)+'.pickle'
+    outfile = open(filename,'wb')
+    pickle.dump(L_speeds,outfile)
+    outfile.close()
+
+    for i in range(len(L_inputs)):
+        filename="models/NEAT/Inputs/"+str(gen).zfill(5)+'_'+str(i).zfill(3)
+        l_inputs=L_inputs[i]
+        outfile = open(filename,'a')
+        for j in range(len(l_inputs)):
+            inputs=l_inputs[j]
+            time=inputs[0]
+            accelerate=inputs[1]
+            brake=inputs[2]
+            steer=inputs[3]
+            if accelerate:
+                outfile.write(str(time)+'press up\n')
+            if brake:
+                outfile.write(str(time)+'press down\n')
+            outfile.write(str(time)+'steer '+str(steer).zfill(5)+'\n')
+        outfile.close()
+
+def run(config_file, checkpoint=None):
+    """
+    Run simulation
+    Parameters
+    ----------
+    config_file: str (path to config file)
+    checkpoint: str (path to checkpoint)
+    Output
+    ----------
+    None
+    """
+    # Load configuration.
+    config = neat.Config(neat.DefaultGenome, neat.DefaultReproduction,
+                         neat.DefaultSpeciesSet, neat.DefaultStagnation,
+                         config_file)
+
+    # Create the population, which is the top-level object for a NEAT run.
+    p = neat.Population(config)
+    if not checkpoint == None:
+        p = neat.Checkpointer.restore_checkpoint(checkpoint)
+
+    # Add a stdout reporter to show progress in the terminal.
+    p.add_reporter(neat.StdOutReporter(True))
+    stats = neat.StatisticsReporter()
+    p.add_reporter(stats)
+    p.add_reporter(neat.Checkpointer(generation_interval=1, filename_prefix=filename_prefix))
+
+    # Run for up to global no generations.
+    winner = p.run(eval_genomes, no_generations)
+    #winner = p.run(eval_genomes, no_generations)
+
+    # Display the winning genome.
+    print('\nBest genome:\n{!s}'.format(winner))
+
+    # Show output of the most fit genome against training data.
+    print('\nOutput:')
+    winner_net = neat.nn.FeedForwardNetwork.create(winner, config)
+
+    node_names = {0:'r', 1:'l', 2:'u'}
+    try:
+        visualize.draw_net(config, winner, True, node_names=node_names)
+        visualize.plot_stats(stats, ylog=False, view=True)
+        visualize.plot_species(stats, view=True)
+    except:
+        print('Missing visualize.py file.')
+
+    # p = neat.Checkpointer.restore_checkpoint('neat-checkpoint-9')
+    # p.run(eval_genomes, 10)
+
+
+def main():
+    """ Main function.
+    Launches the simulation.
+    """
+    local_dir = os.getcwd()
+    config_path = os.path.join(local_dir, 'models/NEAT/config-feedforward')
+    print('Press z to begin.')
+    keyboard.wait('z')
+
+    if checkpoint == None:
+        for cpt in os.listdir('.'):
+            if cpt[:4] == 'neat':
+                os.unlink('./'+cpt)
+
+    run(config_path, checkpoint)
+    
+
+if __name__ == '__main__':
+    main()
