+++ trackmanai/neat_trainer.py
@@ -3,7 +3,6 @@ Functions to use to build a NEAT system.
 """
 
 # Imports
-from glob import glob
 import os
 import keyboard
 import neat
@@ -20,118 +19,134 @@ except ModuleNotFoundError:
     print('Missing visualize.py file.')
 
 
-# Functions
-def eval_genomes(genomes, config):
+# Create NEAT Trainer class
+class NEATTrainer(TMTrainer):
+    """ NEAT TM Trainer.
+    This class has NEAT-related methods :
+        - Genomes evaluation
+        - Training run
     """
-    Evaluates every genome (NN) of the generation
+    def __init__(self, model_dir, model_config, w, h, window_name, 
+    n_lines, server_name, gamespeed, skip_frames, 
+    kill_time, kill_speed, max_time, no_lines, 
+    screen_viewer, checkpoint=None):
+        super().__init__(model_dir, model_config, w, h, window_name, 
+        n_lines, server_name, gamespeed, skip_frames, 
+        kill_time, kill_speed, max_time, no_lines,
+        screen_viewer, checkpoint)
 
-    Parameters
-    ----------
-    genomes: Array(genome) (every NN of the simulation)
-    config: neat.Config (config for the neat algorithm)
 
-    Output
-    ----------
-    None
-    """
-    # Initialize client
-    global gen
-    gen+=1
-    L_net=[]
-    for genome_id, genome in genomes:
-        net = neat.nn.FeedForwardNetwork.create(genome, config)
-        L_net.append(net)
-    max_time2=min(max_time,1+0.5*gen)
+    def eval_genomes(self, genomes, config):
+        """
+        Evaluates every genome (NN) of the generation
 
-    # Run gen
-    L_fit,L_coords,L_speeds,L_inputs=run_client_gen(GenClient(L_net,max_time2,kill_time, sv, skip_frames, window_name, gamespeed, kill_speed),server_name) #1/6 de sec en plus par génération
+        Parameters
+        ----------
+        genomes: Array(genome) (every NN of the simulation)
+        config: neat.Config (config for the neat algorithm)
 
-    # Update fitness
-    for i in range(len(L_fit)):
-        genomes[i][1].fitness=L_fit[i]
+        Output
+        ----------
+        None
+        """
+        # Initialize client
+        self.gen+=1
+        L_net=[]
+        for genome_id, genome in genomes:
+            net = neat.nn.FeedForwardNetwork.create(genome, config)
+            L_net.append(net)
+        max_time2=min(self.max_time,1+0.5*self.gen)
 
-    # Write generation recap
-    filename = 'models/NEAT/Coords/'+str(gen).zfill(5)+'.pickle'
-    outfile = open(filename,'wb')
-    pickle.dump(L_coords,outfile)
-    outfile.close()
-    filename = 'models/NEAT/Speeds/'+str(gen).zfill(5)+'.pickle'
-    outfile = open(filename,'wb')
-    pickle.dump(L_speeds,outfile)
-    outfile.close()
+        # Run gen
+        L_fit,L_coords,L_speeds,L_inputs=run_client_gen(GenClient(L_net,max_time2,self.kill_time, self.screen_viewer, self.skip_frames, self.window_name, self.gamespeed, self.kill_speed),self.server_name) #1/6 de sec en plus par génération
 
-    for i in range(len(L_inputs)):
-        filename="models/NEAT/Inputs/"+str(gen).zfill(5)+'_'+str(i).zfill(3)
-        l_inputs=L_inputs[i]
-        outfile = open(filename,'a')
-        for j in range(len(l_inputs)):
-            inputs=l_inputs[j]
-            time=inputs[0]
-            accelerate=inputs[1]
-            brake=inputs[2]
-            steer=inputs[3]
-            if accelerate:
-                outfile.write(str(time)+'press up\n')
-            if brake:
-                outfile.write(str(time)+'press down\n')
-            outfile.write(str(time)+'steer '+str(steer).zfill(5)+'\n')
+        # Update fitness
+        for i in range(len(L_fit)):
+            genomes[i][1].fitness=L_fit[i]
+
+        # Write generation recap
+        filename = self.model_dir+'/Coords/'+str(self.gen).zfill(5)+'.pickle'
+        outfile = open(filename,'wb')
+        pickle.dump(L_coords,outfile)
+        outfile.close()
+        filename = self.model_dir+'/Speeds/'+str(self.gen).zfill(5)+'.pickle'
+        outfile = open(filename,'wb')
+        pickle.dump(L_speeds,outfile)
         outfile.close()
 
-
-def run(config_file, checkpoint=None, no_generations=1000):
-    """
-    Run simulation
-
-    Parameters
-    ----------
-    config_file: str (path to config file)
-    checkpoint: str (path to checkpoint)
-    no_generations: int (number of generations)
-
-    Output
-    ----------
-    None
-    """
-    # Load configuration.
-    config = neat.Config(neat.DefaultGenome, neat.DefaultReproduction,
-                         neat.DefaultSpeciesSet, neat.DefaultStagnation,
-                         config_file)
-
-    # Create the population, which is the top-level object for a NEAT run.
-    p = neat.Population(config)
-    if not checkpoint == None:
-        p = neat.Checkpointer.restore_checkpoint(checkpoint)
-
-    # Add a stdout reporter to show progress in the terminal.
-    p.add_reporter(neat.StdOutReporter(True))
-    stats = neat.StatisticsReporter()
-    p.add_reporter(stats)
-    p.add_reporter(neat.Checkpointer(generation_interval=1, filename_prefix=filename_prefix))
-
-    # Run for up to global no generations.
-    winner = p.run(eval_genomes, no_generations)
-    #winner = p.run(eval_genomes, no_generations)
-
-    # Display the winning genome.
-    print('\nBest genome:\n{!s}'.format(winner))
-
-    # Show output of the most fit genome against training data.
-    print('\nOutput:')
-    winner_net = neat.nn.FeedForwardNetwork.create(winner, config)
-
-    node_names = {0:'r', 1:'l', 2:'u'}
-    try:
-        visualize.draw_net(config, winner, True, node_names=node_names)
-        visualize.plot_stats(stats, ylog=False, view=True)
-        visualize.plot_species(stats, view=True)
-    except:
-        print('Missing visualize.py file.')
-
-    # p = neat.Checkpointer.restore_checkpoint('neat-checkpoint-9')
-    # p.run(eval_genomes, 10)
+        for i in range(len(L_inputs)):
+            filename=self.model_dir+"/Inputs/"+str(self.gen).zfill(5)+'_'+str(i).zfill(3)
+            l_inputs=L_inputs[i]
+            outfile = open(filename,'a')
+            for j in range(len(l_inputs)):
+                inputs=l_inputs[j]
+                time=inputs[0]
+                accelerate=inputs[1]
+                brake=inputs[2]
+                steer=inputs[3]
+                if accelerate:
+                    outfile.write(str(time)+'press up\n')
+                if brake:
+                    outfile.write(str(time)+'press down\n')
+                outfile.write(str(time)+'steer '+str(steer).zfill(5)+'\n')
+            outfile.close()
 
 
-def train_neat(run_config="./models/config.ini",
+    def run(self, no_generations=1000):
+        """
+        Run simulation
+
+        Parameters
+        ----------
+        config_file: str (path to config file)
+        checkpoint: str (path to checkpoint)
+        no_generations: int (number of generations)
+
+        Output
+        ----------
+        None
+        """
+        # Load configuration.
+        config = neat.Config(neat.DefaultGenome, neat.DefaultReproduction,
+                            neat.DefaultSpeciesSet, neat.DefaultStagnation,
+                            self.model_config)
+
+        # Create the population, which is the top-level object for a NEAT run.
+        p = neat.Population(config)
+        if not self.checkpoint == None:
+            p = neat.Checkpointer.restore_checkpoint(self.checkpoint)
+
+        # Add a stdout reporter to show progress in the terminal.
+        p.add_reporter(neat.StdOutReporter(True))
+        stats = neat.StatisticsReporter()
+        p.add_reporter(stats)
+        p.add_reporter(neat.Checkpointer(generation_interval=1, filename_prefix=self.filename_prefix))
+
+        # Run for up to global no generations.
+        winner = p.run(self.eval_genomes, no_generations)
+        #winner = p.run(eval_genomes, no_generations)
+
+        # Display the winning genome.
+        print('\nBest genome:\n{!s}'.format(winner))
+
+        # Show output of the most fit genome against training data.
+        print('\nOutput:')
+        winner_net = neat.nn.FeedForwardNetwork.create(winner, config)
+
+        node_names = {0:'r', 1:'l', 2:'u'}
+        try:
+            visualize.draw_net(config, winner, True, node_names=node_names)
+            visualize.plot_stats(stats, ylog=False, view=True)
+            visualize.plot_species(stats, view=True)
+        except:
+            print('Missing visualize.py file.')
+
+        # p = neat.Checkpointer.restore_checkpoint('neat-checkpoint-9')
+        # p.run(eval_genomes, 10)
+
+
+def train_neat(model_dir="./models/NEAT",
+    run_config="./models/config.ini",
     model_config="./models/NEAT/config-feedforward",
     checkpoint="./models/NEAT/Checkpoints/checkpoint-0",
     no_generations=1000):
@@ -140,6 +155,7 @@ def train_neat(run_config="./models/config.ini",
 
     Parameters
     ----------
+    model_dir: str (path to model)
     run_config: str (path to run configuration)
     model_config: str (path to model configuration)
     checkpoint: str (name of the checkpoint to load)
@@ -154,30 +170,24 @@ def train_neat(run_config="./models/config.ini",
     config_file.read(run_config)
 
     # Define config variables
-    global w, h, window_name, n_lines, server_name, gamespeed, skip_frames, kill_time, kill_speed, max_time, no_lines
     w, h, window_name = int(config_file['Window']['w']), int(config_file['Window']['h']), config_file['Window']['window_name']
     n_lines = int(config_file['Image']['n_lines'])
     server_name, gamespeed, skip_frames = config_file['Game']['server_name'], int(config_file['Game']['gamespeed']), int(config_file['Game']['skip_frames'])
     kill_time, kill_speed = int(config_file['Game']['kill_time']), int(config_file['Game']['kill_speed'])
     max_time, no_lines = int(config_file['Game']['max_time']), int(config_file['Game']['no_lines'])
-    global sv
-    sv = ScreenViewer(n_lines, w, h)
+    screen_viewer = ScreenViewer(n_lines, w, h)
 
-    # Get information from checkpoint
-    checkpoint_infos = checkpoint.split('/')[-1].split('-')
-    global filename_prefix, gen
-    filename_prefix, gen = checkpoint_infos[0], int(checkpoint_infos[-1])
+    trainer = NEATTrainer(model_dir, model_config, w, h, window_name, 
+    n_lines, server_name, gamespeed, skip_frames,
+    kill_time, kill_speed, max_time, no_lines,
+    screen_viewer, checkpoint)
 
     # Wait for user's input
     print('Press z to begin.')
     keyboard.wait('z')
 
-    if checkpoint == None:
-        for cpt in os.listdir('.'):
-            if cpt[:4] == 'neat':
-                os.unlink('./'+cpt)
-
-    run(model_config, checkpoint, no_generations)
+    # Run training
+    trainer.run(no_generations)
 
 
 if __name__ == '__main__':
